Move enums from gitea to types to avoid circular deps
